Clean up util package.
