+++ digitalocean/util/schema.go
@@ -6,6 +6,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
+// SetResourceDataFromMap sets a *schema.ResourceData from a map.
 func SetResourceDataFromMap(d *schema.ResourceData, m map[string]interface{}) error {
 	for key, value := range m {
 		if err := d.Set(key, value); err != nil {
